Respawning Game without start/stopping SyncManager
Summary: Jank but works only for local play. Needs cleanup and implementation for other connection types.

Problem: Score incrementing after a Fighter death can't be rollbacked.

Info: SyncManager runs continuously starting from when a connection type is selected. SyncManager does not like it when an object is spawned on the frame when it is started for the first time (for some reason). To fix this, a timer of one frame was added to the crude_connection scene. The timer is started after a connection button is pushed, and the Game is spawned after it finishes. A quick solution is implemented to prevent a player's score from incrementing multiple times in the same tick because rollback can cause the score increase to be called more than once at a time.