Basic sound effects
Animations will have keyframes to call a function on Fighter that will play a sound using the rollback netcode framework. The function call has one parameter for a keyword used to reference the sound to be played. Sounds are preloaded in Fighter and stored in a dictionary with a hardcoded (bruh) keyword for each sound. Probably a better way to do this by setting the dictionary in inspector for each character.